test: split apply decision tests
